// src/lib/api/__generated/openapi.ts/zod.gen.ts
export const zCreateLootAwardRequest = z.object({
    characterId: z.string(),
    itemId: z.coerce.bigint().min(BigInt('-9223372036854775808'), { error: 'Invalid value: Expected int64 to be >= -9223372036854775808' }).max(BigInt('9223372036854775807'), { error: 'Invalid value: Expected int64 to be <= 9223372036854775807' }),
    itemName: z.string(),
    itemLevel: z.int().nullish(),
    slot: z.string().nullish(),
    decisionType: z.string(),
    priority: z.string().nullish(),
    note: z.string().nullish()
});

export const zVoidLootAwardRequest = z.object({
    reason: z.string()
});
// ...
export const zCreateLootAwardBody = zCreateLootAwardRequest;

export const zCreateLootAwardPath = z.object({
    raidId: z.string()
});

/**
 * Created
 */
export const zCreateLootAwardResponse = zLootAwardEnvelope;

export const zVoidLootAwardBody = zVoidLootAwardRequest;